scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "mysql-connector-python>=9.2.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "selenium>=4.29.0",
    "webdriver-manager>=4.0.2",
]
